% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_mvtnorm.R
\name{log_mvtnorm}
\alias{log_mvtnorm}
\title{Log of Multivariate Normal Density (Based on Another Package)}
\usage{
log_mvtnorm(x, mu, Sigma, loglik = TRUE)
}
\arguments{
\item{x}{is a vector of sample point or a matrix of sample points}

\item{mu}{is the mean vector of multivariate normal distribution}

\item{Sigma}{is the variance-covariance matrix of multivariate normal distribution}

\item{loglik}{is a logistic argument that determines whether or not to take the
logarithm of the density: \code{TRUE} (default), taking the logarithm; \code{FALSE}, 
otherwise.}
}
\value{
a value (vector or scalar) of the (logarithm of) likelihood of the sample point(s) with the
given multivariate distribution.
}
\description{
This function derives the logarithm of the likelihood of a sample point drawn
from a multivariate normal distribution based on \code{mvtnorm} package.
}
\examples{
set.seed(314159)
mydata <- cbind(c(1, 2), c(3, 4))
mymu <- c(0, 0)
mySigma <- diag(2)
log_mvtnorm(x = mydata, mu = mymu, Sigma = mySigma)

}
\references{
Zhang, P. (2024). Paper name. *Journal name*. **volume**(number), 1--20.
}
\keyword{internal}
