% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logdmvnorm}
\alias{logdmvnorm}
\title{Derive the logarithm of the likelihood of a sample point from a multivariate normal distribution}
\usage{
logdmvnorm(X, mean, Sigma)
}
\arguments{
\item{X}{Vector, a vector of numeric values}

\item{mean}{Vector, the mean of multivariate normal distribution}

\item{Sigma}{Matrix, the variance-covariance matrix of multivariate normal distribution}
}
\value{
Returns the logarithm of the likelihood of X from the multivariate normal distribution 
with mean \code{mean} and variance \code{Sigma}
}
\description{
Derive the logarithm of the likelihood of a sample point from a multivariate normal distribution
}
\keyword{internal}
