% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scsbuildr-package.R
\docType{package}
\name{scsbuildr-package}
\alias{scsbuildr}
\alias{scsbuildr-package}
\title{scsbuildr: Statistical Computing Series: Building an R Package}
\description{
This package provides support to a presentation for the Statistical Computing Serires hosted by the Department of Biostatistics, Vanderbilt University Medical Center. The goal is to demonstrate fundamentals for building an R pacakge.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/panpanzhang99299/scsbuildr}
  \item Report bugs at \url{https://github.com/panpanzhang99299/scsbuildr/issues}
}

}
\author{
\strong{Maintainer}: Panpan Zhang \email{panpan.zhang@vumc.org} (\href{https://orcid.org/0000-0002-8211-5930}{ORCID})

}
\keyword{internal}
